"Enhance CustomTabsManager with custom tabs lifecycle management"
This commit enhances our CustomTabsManager by adding a crucial part of the lifecycle that is the connection/disconnection to the CustomTabsService. The change includes warming up the Browser (to speed up the launching of the URL) and creating new sessions for each URL launching. This was done to improve the performance and user experience when launching URLs in our app. We've also replaced the simple CustomTabsIntent launching with a more flexible approach allowing customization in the future.